[ADD] eLearning: adds eLearning to index and adds eLearning documentation
Added eLearning to the list of applications on the index. Created a toc tree with the various
documentation that will be covered. Purposefully left "Private YouTube videos" off the toc tree, but
still use it as a link in Create Courses. Created documentation for Create Courses, Add Content,
Create Certifications, Sell courses & certifcations, Forums in eLearning, Mailing in eLearning.